requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aioresponses==0.7.6
aiosignal==1.3.1
albumentations==1.4.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
APScheduler==3.10.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.32.59
backoff==2.2.1
blinker==1.7.0
boto3==1.34.59
botocore==1.34.59
branca==0.7.2
brotlipy==0.7.0
cachetools==5.3.3
certifi==2023.7.22
chardet==4.0.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cramjam==2.8.3
cycler==0.10.0
Cython==3.0.0
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker==6.1.3
docutils==0.16
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
fastparquet==2024.2.0
filelock==3.13.1
fiona==1.9.6
flatbuffers==24.3.7
folium==0.16.0
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2024.2.0
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.42
GPUtil==1.4.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
idna==2.10
imageio==2.33.1
imageio-ffmpeg==0.4.9
imutils==0.5.4
inference==0.9.15
inference-cli==0.9.15
inference-sdk==0.9.15
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.25.2
onnxruntime==1.15.1
openai==1.13.3
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
pandas==2.1.4
parso==0.8.3
pendulum==3.0.0
pexpect==4.9.0
piexif==1.1.3
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.22.0
pluggy==1.4.0
prettytable==3.10.0
proglog==0.1.10
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.5.1
pybase64==1.3.2
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pytest==8.0.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.2
referencing==0.33.0
requests-toolbelt==1.0.0
rich==13.5.2
roboflow==1.1.23
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.2
shapely==2.0.1
skypilot==0.4.1
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.16
stack-data==0.6.3
starlette==0.36.3
streamlit==1.32.2
streamlit-drawable-canvas==0.9.3
streamlit-image-comparison==0.0.4
streamlit-image-coordinates==0.1.6
streamlit-metrics==0.1.0
streamlit_folium==0.20.0
structlog==24.1.0
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2024.2.12
time-machine==2.14.0
toml==0.10.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
traitlets==5.14.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
ultralytics==8.1.24
watchdog==4.0.0
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xlrd==2.0.1
yarl==1.9.4
